requirements.txt
pyaudio
numpy
torch
silero-vad
faster-whisper
piper-tts
pydub
python-dotenv
google-genai
fastapi 
uvicorn